+++ Snoop.Core/Windows/ScreenshotDialog.xaml.cs
@@ -19,6 +19,8 @@ namespace Snoop.Windows
     /// </summary>
     public partial class ScreenshotDialog
     {
+        private static string lastSaveDirectory = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
+
         public static readonly RoutedCommand SaveCommand = new RoutedCommand(nameof(SaveCommand), typeof(ScreenshotDialog));
         public static readonly RoutedCommand CancelCommand = new RoutedCommand(nameof(CancelCommand), typeof(ScreenshotDialog));
 
@@ -30,23 +32,6 @@ namespace Snoop.Windows
             this.CommandBindings.Add(new CommandBinding(CancelCommand, this.HandleCancel, (x, y) => y.CanExecute = true));
         }
 
-        #region FilePath Dependency Property
-
-        public string FilePath
-        {
-            get { return (string)this.GetValue(FilePathProperty); }
-            set { this.SetValue(FilePathProperty, value); }
-        }
-
-        public static readonly DependencyProperty FilePathProperty =
-            DependencyProperty.Register(
-                nameof(FilePath),
-                typeof(string),
-                typeof(ScreenshotDialog),
-                new PropertyMetadata());
-
-        #endregion
-
         private void HandleCanSave(object sender, CanExecuteRoutedEventArgs e)
         {
             e.CanExecute = this.DataContext is Visual;
@@ -54,39 +39,40 @@ namespace Snoop.Windows
 
         private void HandleSave(object sender, ExecutedRoutedEventArgs e)
         {
-            if (string.IsNullOrEmpty(this.FilePath))
+            var dpiText = ((TextBlock)((ComboBoxItem)this.dpiBox.SelectedItem).Content).Text;
+
+            var filename = "SnoopScreenshot";
+
+            if (this.DataContext is FrameworkElement element
+                && string.IsNullOrEmpty(element.Name) == false)
             {
-                var filePath = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
-
-                var filename = "SnoopScreenshot.png";
-
-                if (this.DataContext is FrameworkElement element
-                    && string.IsNullOrEmpty(element.Name) == false)
-                {
-                    filename = $"SnoopScreenshot_{element.Name}.png";
-                }
-
-                this.FilePath = Path.Combine(filePath, filename);
+                filename = $"SnoopScreenshot_{element.Name}";
             }
+
+            filename += "_" + dpiText;
+
+            filename += ".png";
+
+            var filePath = Path.Combine(lastSaveDirectory, filename);
 
             var fileDialog = new SaveFileDialog
             {
                 AddExtension = true,
                 CheckPathExists = true,
                 DefaultExt = "png",
-                InitialDirectory = Path.GetDirectoryName(this.FilePath),
-                FileName = Path.GetFileNameWithoutExtension(this.FilePath),
+                InitialDirectory = Path.GetDirectoryName(filePath),
+                FileName = Path.GetFileNameWithoutExtension(filePath),
                 Filter = "Image File (*.png)|*.png",
                 FilterIndex = 0
             };
 
             if (fileDialog.ShowDialog(this).Value)
             {
-                this.FilePath = fileDialog.FileName;
+                lastSaveDirectory = Path.GetDirectoryName(fileDialog.FileName);
 
                 VisualCaptureUtil.SaveVisual(this.DataContext as Visual,
-                    int.Parse(((TextBlock)((ComboBoxItem)this.dpiBox.SelectedItem).Content).Text),
-                    this.FilePath);
+                    int.Parse(dpiText),
+                    filePath);
 
                 this.Close();
             }