+++ Changelog.md
@@ -8,6 +8,7 @@
   - [#45](../../issues/45) - Keystrokes go to Visual Studio main window when inspecting Visual Studio (thanks @KirillOsenkov)
   - [#66](../../issues/66) - System.InvalidOperationException: Collection was modified; enumeration operation may not execute.
   - [#101](../../issues/101) - My Style is being applied to the "Change Target" Window
+  - [#120](../../issues/120) - Screenshot feature produces pixelated low-res image for larger windows
   - [#150](../../issues/150) - Format and parse property values with the same format provider
   - [#151](../../issues/151) - Dependency properties are filtered wrong and less properties are shown than should be
   - [#152](../../issues/152) - Magnified view only works for main window